Treat gitlabUrl as a format string for gitlab-runner
If the gitlabUrl contains the '%' they will be processed using go's fmt
package. https://golang.org/pkg/fmt/

The two args the formater will be passed are the release name, and the
namespace.

This is mostly so we can include the runner chart in our other charts as
a dependency, by passing it a format string instead of a defined
gitlabUrl.