Cleanup unused mattermost config
Also fix the service port for workhorse and node_exporter so they can
actually be reached through the service.

Moved the port configuration in the config file to be at the bottom, as
these are hardcoded to match the container ports.

The services's ports are what can be changed in this setup.